mana work

+++ Assets/Scripts/Ability/Data/HitBox.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 public class HitBox
 {
 	public HitBoxType Type; //rename if needed
-	public int Radius_or_Height;
-	public int Width;
+	public float Radius_or_Height;
+	public float Width;
 	public HitBox(HitBox hitBox)
 	{
 		Type = hitBox.Type;
